requirements.txt
pandas==2.2.2
scikit-learn==1.5.1
fastapi==0.111.0
uvicorn==0.30.1
joblib==1.4.2
openmeteo-requests>=1.7.2,<2.0.0
requests-cache>=1.2.1,<2.0.0
retry-requests>=2.0.0,<3.0.0
numpy>=2.3.3,<3.0.0
xgboost==3.0.5
